--- mlprogram/interpreters/__init__.py
@@ -1,2 +1,0 @@
-from mlprogram.interpreters.interpreter import BatchedState  # noqa
-from mlprogram.interpreters.interpreter import Interpreter  # noqa
