+++ mlprogram/languages/linediff/interpreter.py
@@ -1,5 +1,5 @@
-from mlprogram.interpreters import BatchedState
-from mlprogram.interpreters import Interpreter as BaseInterpreter
+from mlprogram.languages import BatchedState
+from mlprogram.languages import Interpreter as BaseInterpreter
 from mlprogram.languages.linediff import AST
 from mlprogram.languages.linediff import Delta
 from mlprogram.languages.linediff import Insert
