+++ mlprogram/utils/transform/pbe_with_repl.py
@@ -4,8 +4,8 @@ from typing import Generic
 from mlprogram import Environment
 from mlprogram.languages import Token
 from mlprogram.languages import Expander
-from mlprogram.interpreters import BatchedState
-from mlprogram.interpreters import Interpreter
+from mlprogram.languages import BatchedState
+from mlprogram.languages import Interpreter
 
 Code = TypeVar("Code")
 Input = TypeVar("Input")
