+++ test/utils/transform/test_pbe_with_repl.py
@@ -3,7 +3,7 @@ import numpy as np
 from mlprogram import Environment
 from mlprogram.utils.transform.pbe_with_repl import ToEpisode
 from mlprogram.languages import Token
-from mlprogram.interpreters import Interpreter
+from mlprogram.languages import Interpreter
 
 
 class MockInterpreter(Interpreter):
